% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{bernoulli_poly_approx}
\alias{bernoulli_poly_approx}
\title{get approximate polynomial representation of the data y}
\usage{
bernoulli_poly_approx(y, R, k)
}
\description{
get approximate polynomial representation of the data y
}
