% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_kl_normal_poly}
\alias{compute_kl_normal_poly}
\title{Compute KL divergence from normal (quadratic) to polynomial distributions}
\usage{
compute_kl_normal_poly(mu, var, p, normalize = F)
}
\arguments{
\item{mu}{mean of normal distribution}

\item{var}{variance of normal distribution}

\item{p}{vector of polynomial coefficients}

\item{normalize}{whether to normalize the polynomial density--
note: can optimize without normalizing, but can return negative values}
}
\value{
the KL divergence KL[N(mu, var) | p]
}
\description{
Compute KL divergence between normal distribution and a distribution with
polynomial log-density. Faster that `compute_kl_polynomial` because we
can analytically compute moments undert the normal distribution.
}
