% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{compute_moment_polynomial}
\alias{compute_moment_polynomial}
\title{Compute moment}
\usage{
compute_moment_polynomial(p, k, range = c(-20, 20))
}
\arguments{
\item{p}{polynomial coefficients ordered by increasing degree e.g. (p0, ... pM)
Assumes p is normalized!}

\item{k}{the moment to compute}
}
\value{
E[X^k] = \int x^k exp(f(x)) dx
}
\description{
Compute the moments of a distribution with polynomial log-density
}
