% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_moments.R
\name{compute_moments_polynomial_is}
\alias{compute_moments_polynomial_is}
\title{Compute polynomial moments via importance sampling}
\usage{
compute_moments_polynomial_is(p, K, left = -10, right = 10, n = 64)
}
\arguments{
\item{p}{vector of coefficients}

\item{K}{moments to computes}

\item{left}{left integration boundary}

\item{right}{right integration boundary}

\item{n}{number of points for IS}
}
\description{
Compute polynomial moments via importance sampling
}
