% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation2.R
\name{compute_normal_moments}
\alias{compute_normal_moments}
\title{Compute normal moments}
\usage{
compute_normal_moments(mu, var, k)
}
\arguments{
\item{mu}{mean}

\item{var}{variance}

\item{k}{number of moments to copmute}
}
\description{
Compute first `k` moments of a normal distribution (including 0th moment)
}
