% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{compute_normal_moments2}
\alias{compute_normal_moments2}
\title{Compute normal moments}
\usage{
compute_normal_moments2(mu, var, K)
}
\arguments{
\item{mu}{mean parameter of normal distribution}

\item{var}{variance of normal distribution}

\item{K}{number of moments to compute}
}
\value{
a length K vector with the first K (un-centered) moments of N(mu, var)
}
\description{
Recursively compute moments of normal distribution using reccurence relation
M_k = mu M_{k-1} + var * (k-1) * M_{k-2}
}
