% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation2.R
\name{compute_psi_moments}
\alias{compute_psi_moments}
\title{compute k moments for psi = xb, b ~ N(mu, var)}
\usage{
compute_psi_moments(x, mu, var, k)
}
\description{
compute k moments for psi = xb, b ~ N(mu, var)
}
