% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_ops.R
\name{construct_coef_matrix}
\alias{construct_coef_matrix}
\title{Construct coef matrix}
\usage{
construct_coef_matrix(K)
}
\description{
Upper triangular matrix with entires M[j+1, k+1] = (k choose j) k >= 0
}
