% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_ops.R
\name{construct_shift_matrix_generator}
\alias{construct_shift_matrix_generator}
\title{Construct shift matrix generator}
\usage{
construct_shift_matrix_generator(M)
}
\description{
Returns a function for constructing the shift matrix for a polynomial
of degree M
we construct this function to avoid repeated callse to `construct_coef_matrix`
with a fixed coef matrix we can construct the shift matrix by
   1. constructing a toeplitz matrix of the moments "moment matrix"
   2. an elementwise multiplication of the moment matrix and coef matrix
}
