% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_moments.R
\name{logmean}
\alias{logmean}
\title{Log mean}
\usage{
logmean(logw)
}
\arguments{
\item{logw}{= log(w), w a vector}
}
\value{
log(mean(w))
}
\description{
This function computes log(mean(w)) in a numerically stable way
}
