% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{make_approximation}
\alias{make_approximation}
\title{Make approximation}
\usage{
make_approximation(f, left = -4, right = 4, M = 2, plot = F)
}
\arguments{
\item{f}{a 1d function}

\item{left}{boundary of approximation}

\item{right}{boundary of approximation}

\item{M}{degree of the approximation}

\item{plot}{boolean to create a plot of the function}
}
\description{
make polynomial approximation on the interval [left, right]
}
