% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{make_shift_matrix}
\alias{make_shift_matrix}
\title{Make shift matrix}
\usage{
make_shift_matrix(moments)
}
\arguments{
\item{moments}{moments of y}
}
\description{
Generate matrix that maps coefficients of a polynomial
f(x + y) (represented by coefficients p) to coefficients of
f2(x) = E_{p(y)}[f(x+y)]
}
