% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_poly.R
\name{minimize_kl_normal_poly}
\alias{minimize_kl_normal_poly}
\title{Find Gaussian q that minimizes KL[q || p]}
\usage{
minimize_kl_normal_poly(p, tol = 1e-05, max_iter = 10)
}
\arguments{
\item{p}{vector of polynomial coefficients}

\item{tol}{tolerance on change in KL divergence between iterations}

\item{max_iter}{maximum numbr of iterations}
}
\value{
a list(mu=mu, var=var) with parameters of normal distribution
}
\description{
Minimize KL divergence for fixed-form variantional inference
Use a Laplace approximation as initialization, and then
perform coordinate ascent on mean and variance
}
