% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{normalize_polynomial_log_density}
\alias{normalize_polynomial_log_density}
\title{Normalize polynomial log density}
\usage{
normalize_polynomial_log_density(p)
}
\arguments{
\item{p}{polynomial coefficients ordered by increasing degree e.g. (p0, ... pM)}
}
\description{
update intercept so that \int \exp \{f(x) \} dx = 1
computed numerically via quadrature.
where f(x) is a polynomial with coefficients given by argument `p`
}
