% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_moments.R
\name{normalizelogweights}
\alias{normalizelogweights}
\title{Normalize log weights}
\usage{
normalizelogweights(logw)
}
\arguments{
\item{logw}{unnormalized log-probabilities}
}
\value{
normalized probabilities that sum to 1
}
\description{
Convert unnormalized log probabilities into probabilities
}
