% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_poly.R
\name{poly_to_gaussian}
\alias{poly_to_gaussian}
\title{Convert polynomial to Gaussian}
\usage{
poly_to_gaussian(p, poly = F)
}
\arguments{
\item{p}{the coefficients of a polynomial in increasing order p = c(p0, p1, ..., pK)}
}
\value{
a gaussian approximation to p
}
\description{
Convert polynomial to Gaussian
}
