% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_poly.R
\name{poly_to_gaussian_laplace2}
\alias{poly_to_gaussian_laplace2}
\title{Laplace approximation for polynomial density}
\usage{
poly_to_gaussian_laplace2(p, poly = F)
}
\arguments{
\item{p}{vector of polynomial coefficients}

\item{lower}{lower bound to search for mode}

\item{upper}{upper bound to serach for mode}
}
\value{
a list(mu = mu, var=var) for a normal distribution
  representing the Laplace approximation
}
\description{
Make a quadratic approximation to the polynomial log-density at the mode.
This version is a bit slower for high degree polynomials,
but it doesnt require specifying a range to optimize on.
}
