% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_poly.R
\name{poly_to_gaussian_mode}
\alias{poly_to_gaussian_mode}
\title{Polynomial to Gaussian via Laplace approximation}
\usage{
poly_to_gaussian_mode(p, poly = F)
}
\arguments{
\item{p}{the coefficients of a polynomial in increasing order p = c(p0, p1, ..., pK)}

\item{poly}{boolean if true return polynomial representation of gaussian approximation}
}
\value{
a gaussian centered on the mode of exp(p)
}
\description{
convert (unnormalized) polynomial density to gaussian approximation
by taking the mode of the polynomial density and taking a laplace approximation
NOTE: this isn't necessarily the gaussian that minimizes KL(q || p)?
}
