% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polyf}
\alias{polyf}
\title{Make polynomial function}
\usage{
polyf(p)
}
\arguments{
\item{p}{a vector of coefficients in increasing degree e.g. c(c0, c1, ... c_k)}
}
\value{
a function `f` such that f(x)
 evaluates the polynomial with coefficients `p` at x
}
\description{
Take a vector of coefficients and make a vectorized
function that evaluates the polynomial
}
