% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyser.R
\name{polynomial_approximate_ser}
\alias{polynomial_approximate_ser}
\title{Polynomial approximate SER}
\usage{
polynomial_approximate_ser(A, X, prior_variance)
}
\arguments{
\item{A}{a matrix of polynomial coefficients, each row is an observations}

\item{X}{design matrix}

\item{prior_variance}{prior variance of effect size b ~ N(0, prior_variance)}
}
\value{
list summarizing SER posterior
}
\description{
Fit single effect regression with polynomial approximation
The "data" are polynomial coefficients approximating the log-likelihood of
each observation as a function of the (unobserved) linear predictor
}
