% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyser.R
\name{polynomial_approximate_univariate_regression}
\alias{polynomial_approximate_univariate_regression}
\title{Polynomial approximate univariate regression}
\usage{
polynomial_approximate_univariate_regression(
  A,
  x,
  prior_coef = NULL,
  gaussian = T
)
}
\arguments{
\item{A}{an n x M matrix of coefficients--
i-th row gives coefficients for polynomial approximating
log p(y_i | psi) as a function of the linear predictor psi = xb}

\item{x}{a n vector of covariates}

\item{prior_variance}{the prior variance of the effect}
}
\description{
Fit a univariate regression with polynomial approximation to the likelihood
}
