% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{polynomial_update3}
\alias{polynomial_update3}
\title{update q with polynomial approximation of arbitrary degree}
\usage{
polynomial_update3(m, X, prior_p, q)
}
\description{
update q with polynomial approximation of arbitrary degree
}
