% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{shift_polynomial3}
\alias{shift_polynomial3}
\title{Transform coefficients of a polynomial f(x + y) (represented by coefficients p)
to coefficients of f2(x) = E_{p(y)}[f(x+y)]}
\usage{
shift_polynomial3(p, moments)
}
\arguments{
\item{p}{K+1 coefficients of a degree-k polynomial}

\item{moments}{moments of y (including E[y^0] = 1)}
}
\description{
Transform coefficients of a polynomial f(x + y) (represented by coefficients p)
to coefficients of f2(x) = E_{p(y)}[f(x+y)]
}
