% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{bernoulli_poly_approx}
\alias{bernoulli_poly_approx}
\title{Bernoulli + sigmoid polynomial approximation}
\usage{
bernoulli_poly_approx(y, left, right, M)
}
\arguments{
\item{y}{observation}

\item{left}{left boundary or approximation}

\item{right}{boundardy of approximation}

\item{M}{degree of approximation}

\item{center}{midpoint of approximation region}
}
\value{
coefficients for polynomial approximation f(\psi) \approd log p(y | \psi)
}
\description{
Bernoulli + sigmoid polynomial approximation
}
