% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{compute_psi_moments}
\alias{compute_psi_moments}
\title{compute k moments for psi = xb, b ~ N(mu, var)}
\usage{
compute_psi_moments(x, b_moments)
}
\description{
compute k moments for psi = xb, b ~ N(mu, var)
}
