% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R, R/approximation2.R
\name{poly_to_gaussian}
\alias{poly_to_gaussian}
\title{convert (unnormalized) polynomial density to gaussian approximation
p the coefficients of a polynomial in increasing order p = c(p0, p1, ..., pK)}
\usage{
poly_to_gaussian(p)

poly_to_gaussian(p)
}
\description{
convert (unnormalized) polynomial density to gaussian approximation
p the coefficients of a polynomial in increasing order p = c(p0, p1, ..., pK)

convert (unnormalized) polynomial density to gaussian approximation
p the coefficients of a polynomial in increasing order p = c(p0, p1, ..., pK)
}
